Add tests for real-world (like) `query` use.
It turns out I had made a big mistake, and my mocking of things hadn't
caught it. Anyways, this is a pretty simple problem to solve and is now
tested. I tried too hard to reuse my existing functions and it just
proved to be pretty incorrect. The context validator shouldn't be used
to validate query fields.

This fixes the bug for now, but I'll provide a patch for the deeper
problem as soon as I can. Possibly I should simply take this feature
out.

In a nutshell, validators which work on the context object (`this`)
won't work correctly with the `query` functionality. I'm dumbfounded by
the fact I didn't catch this, even though the default `context_type`
validator uses `this`.

[ci skip]